fix width input, separete components
