+++ src/components/Header.tsx
@@ -1,0 +1,55 @@
+import React, { useRef, useEffect } from 'react';
+
+interface HeaderProps {
+  newTitle: string;
+  setNewTitle: (value: string) => void;
+  addTodo: (title: string) => void;
+  toggleAllTodos: () => void;
+  todosLength: number;
+}
+
+export const Header: React.FC<HeaderProps> = ({
+  newTitle,
+  setNewTitle,
+  addTodo,
+  toggleAllTodos,
+  todosLength,
+}) => {
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  useEffect(() => {
+    inputRef.current?.focus();
+  }, [todosLength]);
+
+  const handleSubmit = (event: React.FormEvent) => {
+    event.preventDefault();
+    if (newTitle.trim()) {
+      addTodo(newTitle);
+      setNewTitle('');
+    }
+  };
+
+  return (
+    <header className="todoapp__header">
+      {todosLength > 0 && (
+        <button
+          data-cy="ToggleAllButton"
+          type="button"
+          className="todoapp__toggle-all"
+          onClick={toggleAllTodos}
+        ></button>
+      )}
+      <form onSubmit={handleSubmit}>
+        <input
+          ref={inputRef}
+          data-cy="NewTodoField"
+          type="text"
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+          value={newTitle}
+          onChange={e => setNewTitle(e.target.value)}
+        />
+      </form>
+    </header>
+  );
+};
