+++ src/styles/todoapp.scss
@@ -25,6 +25,7 @@
 
   &__header {
     position: relative;
+    box-sizing: border-box;
   }
 
   &__toggle-all {
@@ -58,6 +59,7 @@
   &__new-todo {
     width: 100%;
     padding: 16px 16px 16px 60px;
+    box-sizing: border-box;
 
     font-size: 24px;
     line-height: 1.4em;
